Remove node overlaps
This is an improved version of _reduce_node_overlaps.  Instead of
using a heuristic to move nodes toward empty space to move them apart
from one another, this method uses the node size to check if an
overlap exists. If that is the case, the nodes in question are
moved apart until they no longer overlap.